fixed CSV reader

+++ steven.cc
@@ -1,36 +1,36 @@
 #include "steven.h"
 
-Person::Person(std::string n, float e)
+Person::Person(std::string n, int e)
 :name{n}, expence{e}, debt{0}
 //  ----------------------
 {}
 
-Bank::Bank(float tot)
+Bank::Bank(int tot)
 :tot_expences{tot}
 //  ----------------------
 {}
 
-float Bank::get_expences()
+int Bank::get_expences()
 //  ---------------------
 {
     return tot_expences;
 }
-void Person::update_debt(float cost)
+void Person::update_debt(int cost)
 //  ----------------------
 {
     debt -= cost;
 }
-float Person::get_debt() const
+int Person::get_debt() const
 //  ----------------------
 {
     return debt;
 }
-float Person::get_expence() const
+int Person::get_expence() const
 //  ----------------------
 {
     return expence;
 }
-void Person::update_expence(float cost)
+void Person::update_expence(int cost)
 //  ----------------------
 {
 expence += cost;
